backend/requirements.txt
psycopg2==2.8.3
alembic==1.0.11
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Dance==2.2.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Mako==1.0.14
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
oauthlib==3.0.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1.2
redis==3.3.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.1
sqlit==0.1.6
urllib3==1.25.3
URLObject==2.4.3
vine==1.3.0
Werkzeug==0.15.5
